用户管理和角色管理功能，已完成。

+++ src/main/java/com/cqgy/park/domain/SysUserAuthoritys.java
@@ -23,7 +23,7 @@ public class SysUserAuthoritys {
 	private Long UserId;
 	private Long authorityId;
 	
-	private Integer createUser;
+	private Long createUser;
 	@Temporal(TemporalType.TIMESTAMP)
 	private Date createTime;
 	public Long getId() {
@@ -44,10 +44,10 @@ public class SysUserAuthoritys {
 	public void setAuthorityId(Long authorityId) {
 		this.authorityId = authorityId;
 	}
-	public Integer getCreateUser() {
+	public Long getCreateUser() {
 		return createUser;
 	}
-	public void setCreateUser(Integer createUser) {
+	public void setCreateUser(Long createUser) {
 		this.createUser = createUser;
 	}
 	public Date getCreateTime() {
@@ -56,7 +56,7 @@ public class SysUserAuthoritys {
 	public void setCreateTime(Date createTime) {
 		this.createTime = createTime;
 	}	
-	public SysUserAuthoritys(Long id, Long userId, Long authorityId, Integer createUser, Date createTime) {
+	public SysUserAuthoritys(Long id, Long userId, Long authorityId, Long createUser, Date createTime) {
 		super();
 		Id = id;
 		UserId = userId;
