+++ src/main/java/com/cqgy/park/web/SysAuthorityController.java
@@ -110,7 +110,7 @@ public class SysAuthorityController {
 		if(id == 0){
 			sysAuthority.setId(null);
 			sysAuthority.setCreateTime(new Date());
-			sysAuthority.setCreateUser((Integer)session.getAttribute("login_id"));
+			sysAuthority.setCreateUser((Long)session.getAttribute("login_id"));
 		}
 		sysAuthority.setFatherId(father_id);
 		sysAuthority.setFlag(flag);
@@ -120,7 +120,7 @@ public class SysAuthorityController {
 		sysAuthority.setRemark(remark);
 		sysAuthority.setUri(uri);
 		sysAuthority.setUpdateTime(new Date());
-		sysAuthority.setUpdateUser((Integer)session.getAttribute("login_id"));
+		sysAuthority.setUpdateUser((Long)session.getAttribute("login_id"));
 		sysAuthorityRepository.save(sysAuthority);
 		model.addAttribute("result", "创建菜单成功！");
 		String forword="/display/result";
